% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addXICs.r
\name{addXICs}
\alias{addXICs}
\title{Add extracted ion chromatograms (EIC, BPC) to alignment results}
\usage{
addXICs(
  x,
  in_column = "s",
  rt_column = c("average_rt_min", "rt_min"),
  out_column = "bpc",
  xraw,
  mz_dev = 0.01,
  rt_dev = 10,
  zeroVal = NA,
  smooth = 3,
  max_mz = 10,
  EIC = FALSE,
  .pivot_longer = FALSE
)
}
\arguments{
\item{x}{data.frame as created by \code{\link{loadAlignmentResults}()}}

\item{in_column}{name of column containing m/z's to extract: either a list of
spectra or single m/z values}

\item{rt_column}{name of column containing retention times}

\item{out_column}{name of column where results are}

\item{xraw}{list of 'xcmsRaw' objects}

\item{mz_dev}{width of m/z window to extract (passed to
\code{\link{getXIC}()})}

\item{rt_dev}{width of RT window to extract (passed to
\code{\link{getXIC}()})}

\item{zeroVal}{value to replace NA's by (passed to \code{\link{getXIC}()})}

\item{smooth}{(passed to \code{\link{getXIC}()})}

\item{max_mz}{restrict BPC extraction to the 'max_mz' highest peaks (if
'in_column' contains spectra)}

\item{EIC}{if 'TRUE' return (sum-based) EIC instead of (maximum-based) BPC}

\item{.pivot_longer}{transform resulting BPC matrices to long format. Useful
for plotting with 'ggplot'.}
}
\value{
data.frame
}
\description{
Add extracted ion chromatograms (EIC, BPC) to alignment results
}
\examples{
# see ?loadAlignmentResults

}
