% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{checkArgs}
\alias{checkArgs}
\title{check if arg is present in arg list, if not, set it to 'val'}
\usage{
checkArgs(args, name, val)
}
\arguments{
\item{args}{arg list}

\item{name}{arg}

\item{val}{default value}
}
\value{
list
}
\description{
(internal function)
}
