% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSynonyms.r
\name{filterSynonyms}
\alias{filterSynonyms}
\title{Filter chemical synonyms as obtained from querying Pubchem}
\usage{
filterSynonyms(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Filter chemical synonyms as obtained from querying Pubchem
}
