% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_additions.r
\name{filterXcmsRaw}
\alias{filterXcmsRaw}
\title{Intensity filter for 'xcmsRaw'}
\usage{
filterXcmsRaw(xraw, minabs = 0, verbose = TRUE)
}
\arguments{
\item{xraw}{'xcmsRaw' object}

\item{minabs}{intensity threshold}

\item{verbose}{print diagnostic message}
}
\value{
'xcmsRaw'
}
\description{
Internal function used by \code{\link{loadRaws}()}
}
\examples{
# see \code{\link{loadRaws}()}
}
