% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{fml2iso}
\alias{fml2iso}
\title{Calculate isotope pattern using 'enviPat'}
\usage{
fml2iso(fml, resolution = 20000, ...)
}
\arguments{
\item{fml}{chemical formula}

\item{resolution}{mass spectral resolution (m/delta_m)}

\item{...}{passed to 'enviPat::isowrap()'}
}
\value{
list
}
\description{
Calculate isotope pattern using 'enviPat'
}
\examples{
\dontrun{
fml2iso("C6H6O")
}
}
