% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{fml2tbl}
\alias{fml2tbl}
\title{Tabulate elements in chemical formula}
\usage{
fml2tbl(fml, elements = NULL)
}
\arguments{
\item{fml}{chemical formula}

\item{elements}{chemical elements to consider, or NULL for all elements
occurring in 'fml'}
}
\value{
data.frame
}
\description{
(internal function)
}
\examples{
\dontrun{
fml2tbl("C6H6O")
fml2tbl("C6H6O", elements = c("C", "H", "O", "S"))
}

}
