% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_additions.r
\name{getScanAtRT}
\alias{getScanAtRT}
\title{Get scan data for a given retention time}
\usage{
getScanAtRT(xraw, rt, ...)
}
\arguments{
\item{xraw}{'xcmsRaw' object or list of 'xcmsRaw' objects}

\item{rt}{retention time in seconds}

\item{...}{passed to \code{\link[xcms]{getScan}()}}
}
\value{
data.frame
}
\description{
Wrapper for \code{\link[xcms]{getScan}()} accepting a list of \code{xcmsRaw}'s
}
\examples{
 mzml_files <- list.files(system.file("extdata", package = "msdialhelpers"),
"GCQTOF_PAH.*\\\\.mzML", full.names = TRUE)
xraw <- loadRaws(mzml_files)
rt_min <- 21
scan <- getScanAtRT(xraw, rt_min * 60)
plot(scan[[1]], type = "h")
}
