% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXIC.r
\name{getXIC}
\alias{getXIC}
\title{Extract ion chromatograms (BPC, EIC, TIC)}
\usage{
getXIC(
  xraw,
  mz = NULL,
  mz_dev = 0.01,
  rt = NULL,
  rt_dev = 2,
  zeroVal = NA,
  smooth = 0,
  EIC = FALSE
)
}
\arguments{
\item{xraw}{'xcmsRaw' object or list of 'xcmsRaw' objects}

\item{mz}{mass vector or NULL to extract the full mass range}

\item{mz_dev}{allowed m/z deviations, see
\code{\link[HiResTEC]{getMultipleBPC}()} for details}

\item{rt}{retention time point in seconds or NULL to extract the full RT
range}

\item{rt_dev}{RT window}

\item{zeroVal}{value to use for intensities <= 0 (typically NA or 0)}

\item{smooth}{window size for moving average smoother, 0 = no smoothing}

\item{EIC}{if 'TRUE' return (sum-based) EIC instead of (maximum-based) BPC}
}
\value{
A matrix with scan wise (rows) intensities for all requested m/z's
(columns), or a list of such matrices.
}
\description{
Wrapper for \code{\link[HiResTEC]{getMultipleBPC}()} accepting lists of
'xcmsRaw' objects.
}
\examples{
mzml_files <- list.files(system.file("extdata", package = "msdialhelpers"),
"GCQTOF_PAH.*\\\\.mzML", full.names = TRUE)
xraw <- loadRaws(mzml_files)

tic <- getXIC(xraw, EIC = TRUE) # total ion chromatogram (TIC)
plotXIC(tic)

bpc <- getXIC(xraw, EIC = FALSE) # full-range base peak chromatogram (BPC)
plotXIC(bpc)

xic <- getXIC(xraw, mz = c(202.078, 101.042), mz_dev = 0.005, rt = 20*60, rt_dev = 1*60)
plotXIC(xic)

par(mfrow = c(1,2))
plotXIC(xic, set.mfrow = FALSE) # override default multi-figure layout

xic_scaled <- scaleXIC(xic, scope = "by_file")
plotXIC(xic_scaled)
}
