% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{ion2rule}
\alias{ion2rule}
\title{ion2rule}
\usage{
ion2rule(ions = "[M+H]+")
}
\arguments{
\item{ions}{character vector, e.g. "[M+H]+"}
}
\value{
data.frame
}
\description{
(internal function)
}
\examples{
\dontrun{
ion2rule("[M+H]+")
}
}
