% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{isInstalled}
\alias{isInstalled}
\title{Check if a package is installed}
\usage{
isInstalled(pkg, libPaths = .libPaths())
}
\arguments{
\item{pkg}{package to check (character(1))}

\item{libPaths}{library paths to check, defaults to \code{.libPaths()}}
}
\value{
logical
}
\description{
(internal function)
}
