% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_nomenclature.r
\docType{data}
\name{lipid_nomenclature}
\alias{lipid_nomenclature}
\title{Lipid nomenclature used by MS-DIAL}
\format{
data.frame
}
\usage{
data(lipid_nomenclature)
}
\description{
Data.frame giving details for lipid class abbreviations occurring in MS-DIAL results.
}
\keyword{datasets}
