% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAlkanes.r
\name{loadAlkanes}
\alias{loadAlkanes}
\title{Import alkane definition files}
\usage{
loadAlkanes(txtfile, plot = FALSE)
}
\arguments{
\item{txtfile}{text file with two tab-separated columns (Num|RT(min))}

\item{plot}{show diagnostic plot?}
}
\value{
data.frame
}
\description{
Import alkane definition files
}
\examples{
fp <- system.file("extdata/Alkane.txt", package = "msdialhelpers")
loadAlkanes(fp, plot = TRUE)
}
