% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConsoleResults.r
\name{loadConsoleResults}
\alias{loadConsoleResults}
\title{Import MS-DIAL console results}
\usage{
loadConsoleResults(txtfile)
}
\arguments{
\item{txtfile}{.msdial file for individual MS files. Use
\code{\link{loadAlignmentResults}()} to import alignment results for
multiple files.}
}
\value{
data.frame
}
\description{
Import '.msdial' files created by MS-DIAL console app.
}
\examples{
fp <- system.file("extdata/MSDconsole_single_result.msdial", package = "msdialhelpers")
dat <- loadConsoleResults(fp)
}
