% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_additions.r
\name{loadXcmsRaw}
\alias{loadXcmsRaw}
\title{Load MS raw data (single file)}
\usage{
loadXcmsRaw(msfile, profstep = 0, minabs = 0, ...)
}
\arguments{
\item{msfile}{mz(X)ML files to load}

\item{profstep}{profile matrix step size parameter used by \code{\link[xcms]{xcmsRaw}()}}

\item{minabs}{absolute intensity threshold}

\item{...}{further arguments passed to \code{\link[xcms]{xcmsRaw}()}}
}
\value{
'xcmsRaw'
}
\description{
Internal function used by \code{\link{loadRaws}()}
}
\examples{
# see \code{\link{loadRaws}()}
}
