% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNISTsearch.r
\name{runNISTsearch}
\alias{runNISTsearch}
\title{Annotate MS-DIAL spectra using NIST search}
\usage{
runNISTsearch(
  x,
  in_column = "s",
  ri_column = "average_ri",
  out_column = "NISTres",
  ...
)
}
\arguments{
\item{x}{MS-DIAL data.frame as returned by \code{\link{loadConsoleResults}()} or \code{\link{loadAlignmentResults}()}}

\item{in_column}{name of column containing prepared spectra (see \code{\link{prepareSpectra}()})}

\item{ri_column}{name of column containing RI values or 'NULL' to exclude RI}

\item{out_column}{name of column added to data.frame}

\item{...}{further arguments passed to \code{\link{searchNIST}()}}
}
\value{
data.frame
}
\description{
Annotate MS-DIAL alignment results using MSPepSearch
}
