% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleXIC.r
\name{scaleXIC}
\alias{scaleXIC}
\title{Scale chromatograms}
\usage{
scaleXIC(xic, scope = c("global", "by_file", "by_trace", "none")[1])
}
\arguments{
\item{xic}{chromatogram(s) as obtained by \code{\link{getXIC}()}}

\item{scope}{one of "global" (default), "by_file", "by_trace" or "none"}
}
\value{
matrix or list equivalent to original 'xic'
}
\description{
Scale chromatograms
}
\examples{
# see getXIC()
}
