% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNIST.r
\name{searchNIST}
\alias{searchNIST}
\title{Send mass spectrum to MSPepSearch}
\usage{
searchNIST(
  spec,
  searchtype = "I",
  presearchtype = "d",
  hits = 5,
  extralibs = c(""),
  ri_obs = NULL,
  max_ri_dev = 100,
  ri_minscore = 0.9,
  mzlimits = c(-1, -1),
  localRdata = getOption("msdialhelpers_localNISTrda"),
  .debug = FALSE
)
}
\arguments{
\item{spec}{mass spectrum: matrix or dataframe with m/z and intensity in the
first two columns, or a list of matrices}

\item{searchtype}{one of "I" (identity, default), "Q" (quick identity), "S"
(simple similarity), "H" (hybrid similarity), "L" (neutral loss
similarity), "M" (ms/ms in EI library), "P" (peptide ms/ms search in a
ms/ms library (use peak annotations and weighting), "G" (generic ms/ms
search in a ms/ms library (no peak annotations or weighting used))}

\item{presearchtype}{one of "d" (standard (use pre-indexed peaks); default),
"s" (sequential (compare to all library spectra)), "f" (fast presearch (use
pre-indexed peaks)), "m" (precursor ion m/z within precursor ion
uncertainty - only for MS/MS search (P or G))}

\item{hits}{how many hits to return}

\item{extralibs}{libraries to use in addition to 'mainlib' and 'replib';
check "MSSEARCH" directory for further libraries that may be installed}

\item{ri_obs}{measured retention index (RI)}

\item{max_ri_dev}{passed to \code{scoreRI()}}

\item{ri_minscore}{minimum RI score; passed to \code{scoreRI()}}

\item{mzlimits}{set m/z limits for spectral comparison}

\item{localRdata}{attach meta data from local .rda (path to .rda or object
already loaded in workspace)}

\item{.debug}{if TRUE keep temporary files created by MSPepSearch}
}
\value{
data.frame
}
\description{
Use "MSPepSearch64.exe" to search NIST library. Download zip file from NIST
site, copy all files to "MSSEARCH" directory and add the following line to
your .Rprofile: \code{options(msdialhelpers_mspepsearch_path =
"/path/to/MSPepSearch64.exe")}
}
\details{
Tested with MSPepSearch64 v2019_02_22.
}
\examples{
\dontrun{
spec <- str2spec("135:999 107:275 150:262 41:150 95:142 136:103 77:84 39:74 91:72 65:49")
searchNIST(spec)
searchNIST(spec, ri_obs = 1300)
}
}
