% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPubchem.r
\name{searchPubchem}
\alias{searchPubchem}
\title{Add PubChem metadata to MS-DIAL alignment results}
\usage{
searchPubchem(
  x,
  in_column = c("inchikey", "NISTres")[1],
  cache_file = "~/pubchem_cache.rda"
)
}
\arguments{
\item{x}{data.frame}

\item{in_column}{name of column containing InChIKey's. Can be "NISTres" in
which case query is performed on a previous NIST search result.}

\item{cache_file}{full path of results cache file or 'NULL' to disable
caching}
}
\value{
data.frame
}
\description{
Query Pubchem for compound metadata using the webchem package. A cache file
is used to store obtained results for subsequent queries. The function (1)
translates supplied InChIKeys to Pubchem CIDs and (2) queries PubChem for
compound details for these CIDs. Currently, PubChem 'Synonyms', 'Properties',
'Uses' and 'Toxicity' sections are retrieved.
}
\examples{
\dontrun{
ik <- data.frame(inchikey = "MXWJVTOOROXGIU-UHFFFAOYSA-N")
searchPubchem(ik)
}
}
