% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec2str.r
\name{str2spec}
\alias{str2spec}
\title{Convert mass spectrum from character to matrix representation}
\usage{
str2spec(x, colnames = c("mz", "i"))
}
\arguments{
\item{x}{character string encoding mass spectra, e.g. "100:999 101:121"}

\item{colnames}{column names in resulting data.frame (default \code{c("mz", "i")})}
}
\value{
data.frame
}
\description{
Convert mass spectrum from character to matrix representation
}
\examples{
str2spec("100:999 101:121")
}
