% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMSP.r
\name{writeMSP}
\alias{writeMSP}
\title{Export an MSP library file}
\usage{
writeMSP(x, file, append = FALSE, progress = FALSE)
}
\arguments{
\item{x}{data.frame to be exported as MSP}

\item{file}{MSP file to be written}

\item{append}{append to existing file?}

\item{progress}{display progress bar?}
}
\value{
None
}
\description{
Export an MSP library file
}
\examples{
msp <- data.frame(
  Name = "Test_spectrum",
  `Num peaks` = 3,
  PEAKS = "100:999 101:122 102:33",
  check.names = FALSE
)
msp
\dontrun{
writeMSP(msp, file = "test.msp")
}
}
