% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAlignmentResults.r
\name{loadAlignmentResults}
\alias{loadAlignmentResults}
\title{Import alignment results from MS-DIAL}
\usage{
loadAlignmentResults(
  txt,
  raw_folder = NULL,
  raw_suffix = "mzml",
  sample_names = NULL
)
}
\arguments{
\item{txt}{Alignment results file ("Height.txt" or "Area.txt")}

\item{raw_folder}{Location of raw data. If given, file paths of raw data will be attached to \code{attr(x, "msdial_sam")}.}

\item{raw_suffix}{File type of raw data. Adjust this when file type differs from 'mzML'. Only relevant when 'raw_folder' is given.}

\item{sample_names}{Optional character vector to use instead of original column names. No check for proper order is performed.}
}
\value{
data.frame with two attributes attached: "msdial_sam" and "intensity_columns".

\code{attr(x, "msdial_sam")} is derived from the file header and contains sample class information as defined in MS-DIAL.

\code{attr(x, "intensity_columns")} is an index of columns holding MS intensities (sample columns).
}
\description{
Read "Height.txt" or "Area.txt" files obtained by running "Export -> Alignment result" in MS-DIAL. Supports file format created by MS-DIAL versions >= 4.0.
}
\examples{
fp <- system.file("extdata/MSDIAL_Alignment_result_GC-MS.txt", package = "msdialhelpers")
aligned <- loadAlignmentResults(fp)
head(aligned)
}
