% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSDconsole.r
\name{runMSDconsole}
\alias{runMSDconsole}
\title{Trigger MS-DIAL console process}
\usage{
runMSDconsole(
  msdir,
  analysis_type = c("gcms", "lcmsdda", "lcmsdia", "lcimmsdda", "lcimmsdia")[1],
  method = createMSDCmethod(),
  MSDCexe = getOption("msdialhelpers_msdialconsole_path"),
  readResults = TRUE,
  skipRun = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{msdir}{directory holding MS files supported by MS-DIAL (.abf, .mzML, .mzXML, .cdf etc.)}

\item{analysis_type}{one of "gcms", "lcmsdda", "lcmsdia", "lcimmsdda", "lcimmsdia"}

\item{method}{list of parameters as returned by \code{\link{createMSDCmethod}()} or name of MS-DIAL method file (.txt)}

\item{MSDCexe}{full path of MS-DIAL console executable (MsdialConsoleApp.exe)}

\item{readResults}{read .msdial using \code{\link{loadConsoleResults}()} or \code{\link{loadAlignmentResults}()}}

\item{skipRun}{don't run MS-DIAL console, just read existing .msdial files}

\item{verbose}{show output of MS-DIAL console}
}
\value{
list of data.frame's containing MS-DIAL results or vector of .msdial files, depending on 'readResults'
}
\description{
Trigger MS-DIAL console process
}
\examples{
\dontrun{
mzml_files <- list.files(system.file("extdata", package = "msdialhelpers"), 
  "GC-QTOF.*\\\\.mzML", 
  full.names = TRUE)
msdir <- "~/MSDCtest"
dir.create(msdir)
file.copy(mzml_files, msdir)
(resfiles <- runMSDconsole(msdir))
peaks <- loadConsoleResults(resfiles[2])
aligned <- loadAlignmentResults(resfiles[1])
}

## use a custom method file
\dontrun{
(resfiles <- runMSDconsole(msdir, method = createMSDCmethod(minimum_peak_height = 5000)))
peaks <- loadConsoleResults(resfiles[2])
}
}
